requirements.txt
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.7.0
langchain>=0.0.200
openai>=0.27.0
fastapi>=0.95.0
uvicorn>=0.21.0
numpy>=1.21.0
scipy>=1.7.0
tqdm>=4.65.0
datasets>=2.12.0
evaluate>=0.4.0